requirements/dev.txt
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
psycopg2==2.9.3
WTForms==3.0.1